+++ src/mash/shell/__init__.py
@@ -1,4 +1,4 @@
-from mash.shell.base import ShellError, ShellPipeError, ShellSyntaxError
+from mash.shell.errors import ShellError, ShellPipeError, ShellSyntaxError
 from mash.shell.shell import Shell
 from mash.shell.function import ShellFunction
 from mash.shell.with_filesystem import ShellWithFileSystem
