+++ src/mash/shell/base.py
@@ -1,130 +1,33 @@
-from asyncio import CancelledError
-import cmd
+from collections import defaultdict
 from dataclasses import asdict
-from itertools import repeat
 from json import dumps, loads
-from typing import Any, Callable, Dict, Iterable, List, Tuple, Union
+from typing import Any, Callable, Dict, List
 import logging
-import subprocess
 
-from mash import io_util
+from mash.shell.cmd2 import Cmd2
 from mash.filesystem.filesystem import FileSystem
 from mash.filesystem.scope import Scope, show
-from mash.io_util import log, shell_ready_signal, print_shell_ready_signal, check_output
+from mash.io_util import log, print_shell_ready_signal
 from mash.shell import delimiters
-from mash.shell.delimiters import DEFINE_FUNCTION, FALSE, IF, TRUE
-from mash.shell.errors import ShellError, ShellPipeError, ShellSyntaxError
-from mash.shell.function import InlineFunction
-from mash.shell.if_statement import Abort,  handle_prev_then_else_statements
-from mash.shell.lex_parser import parse
-from mash.shell.model import LAST_RESULTS, LAST_RESULTS_INDEX, ElseCondition, Indent, Lines, Map, Node, Term, Terms, enter_new_scope, scope
-from mash.shell.parsing import filter_comments, quote_items, to_bool
-from mash.util import has_method, identity, is_valid_method_name, quote_all
+from mash.shell.delimiters import FALSE, IF, TRUE
+from mash.shell.errors import ShellError
+from mash.shell.function import LAST_RESULTS, LAST_RESULTS_INDEX, InlineFunction, scope
+from mash.shell.function import ShellFunction as Function
+from mash.shell.parsing import to_bool
+from mash.util import has_method, identity
 
 
-confirmation_mode = False
 default_session_filename = '.shell_session.json'
-default_prompt = '$ '
+default_function_group_key = '_'
 
-COMMENT = '#'
-RAW_LINE_INDENT = 'raw_line_indent'
 ENV = 'env'
+CHAR = 'char'
 
-Command = Callable[[cmd.Cmd, str], str]
-Types = Union[str, bool, int, float]
-
-
-class Cmd(cmd.Cmd):
-    """Extend CMD with various capabilities.
-    This class is restricted to functionality that requires Cmd methods to be overrriden.
-
-    Features:
-    - Confirmation mode to allow a user to accept or decline commands.
-    - Error handling.
-    """
-
-    intro = 'Press ctrl-d to exit, ctrl-c to cancel, ? for help, ! for shell interop.\n' + \
-        shell_ready_signal + '\n'
-
-    prompt = default_prompt
-    completenames_options = []
-
-    def onecmd(self, lines: str) -> bool:
-        """Parse and run `line`.
-        Returns 0 on success and None otherwise
-        """
-        if lines == 'EOF':
-            logging.debug('Aborting: received EOF')
-            exit()
-
-        try:
-            lines = self.onecmd_prehook(lines)
-            return self.onecmd_inner(lines)
-
-        except ShellSyntaxError as e:
-            if self.ignore_invalid_syntax:
-                log(e)
-            else:
-                raise
-
-        except CancelledError:
-            pass
-
-    def onecmd_inner(self, lines: str):
-        return super().onecmd(lines)
-
-    def onecmd_prehook(self, line):
-        """Similar to cmd.precmd but executed before cmd.onecmd
-        """
-        if confirmation_mode:
-            assert io_util.interactive
-            log('Command:', line)
-            if not io_util.confirm():
-                raise CancelledError()
-
-        return line
-
-    def pipe_cmd_py(self, line: str, result: str):
-        # append arguments
-        line = f'{line} {result}'
-
-        return super().onecmd(line)
-
-    def completenames(self, text, *ignored):
-        """Conditionally override Cmd.completenames
-        """
-        if self.completenames_options:
-            return [a for a in self.completenames_options if a.startswith(text)]
-
-        return super().completenames(text, *ignored)
-
-    def default(self, line: str):
-        # TODO move this
-        if self.ignore_invalid_syntax:
-            return super().default(line)
-
-        raise ShellSyntaxError(f'Unknown syntax: {line}')
-
-    def none(self, _: str) -> str:
-        """Do nothing. Similar to util.none.
-        """
-        return ''
-
-    ############################################################################
-    # Commands: do_*
-    ############################################################################
-
-    def do_shell(self, args):
-        """System call
-        """
-        logging.info(f'Cmd = !{args}')
-        return check_output(args)
-
-    def do_fail(self, msg: str):
-        raise ShellError(f'Fail: {msg}')
+Command = Callable[[Cmd2, str], str]
+FunctionGroup = Dict[str, Dict[str, Function]]
 
 
-class BaseShell(Cmd):
+class BaseShell(Cmd2):
     """Extend CMD with various capabilities.
     This class is restricted to functionality that requires Cmd methods to be overrride.
 
@@ -178,8 +81,9 @@ class BaseShell(Cmd):
         self.auto_save = False
         self.auto_reload = False
 
+        self.function_groups: FunctionGroup = defaultdict(dict)
+
         # internals
-        self._char_methods = {}
         self._default_method = identity
 
     def init_current_scope(self):
@@ -207,15 +111,6 @@ class BaseShell(Cmd):
     def _last_results_index(self):
         return self.env[LAST_RESULTS_INDEX]
 
-    def set_special_method(self, char: str, method: Command):
-        if char in delimiters.all or char in '!?':
-            raise ShellError(f'Char {char} is already in use.')
-
-        self._char_methods[char] = method
-
-    def run_special_method(self, k: str, *args):
-        return self._char_methods[k](*args)
-
     def _save_result(self, value):
         logging.debug(f'_save_result [{self._last_results_index}]: {value}')
 
@@ -226,16 +121,21 @@ class BaseShell(Cmd):
 
         self._last_results[self._last_results_index] = value
 
-    def is_special_method(self, char: str) -> bool:
-        """Check whether `char` is a special characters method. 
+    def is_hidden_function(self, k: str) -> bool:
+        """Check whether `k` is an existing function.
         """
-        return char in self._char_methods
+        return any(k in keys for keys in self.function_groups.values())
+
+    def is_special_function(self, k: str) -> bool:
+        """Check whether `char` is a special characters method
+        """
+        return k in self.function_groups[CHAR]
 
     def is_function(self, k: str) -> bool:
         """Check whether `k` is an existing function. 
         """
         return has_method(self, f'do_{k}') \
-            or self.is_special_method(k) \
+            or self.is_hidden_function(k) \
             or self.is_inline_function(k) \
             or k in '!?'
 
@@ -243,6 +143,9 @@ class BaseShell(Cmd):
         """Check whether `k` is an existing inline (user-defined) function. 
         """
         return k in self.env and isinstance(self.env[k], InlineFunction)
+
+    def run_commands(self, ast: str, prev_result='', run=False):
+        raise NotImplementedError()
 
     ############################################################################
     # Commands: do_*
@@ -260,52 +163,6 @@ class BaseShell(Cmd):
 
         return ''
 
-    def do_map(self, args=''):
-        """Apply a function to every line.
-        If `$` is present, then each line from stdin is inserted there.
-        Otherwise each line is appended.
-
-        Usage
-        -----
-        ```sh
-        println a b |> map echo
-        println a b |> map echo prefix $ suffix
-        ```
-        """
-        log('Expected arguments')
-        return ''
-
-    def _do_foreach(self, ast: tuple, prev_results: str) -> Iterable:
-        """Apply a function to every term or word.
-
-        Usage
-        ```sh
-        echo a b |> foreach echo
-        echo a b |> foreach echo prefix $ suffix
-        ```
-        """
-        prev_results = '\n'.join(prev_results.split(' '))
-        return Map.map(ast, prev_results, self, delimiter=' ')
-
-    def foldr(self, commands: List[Term], prev_results: str, delimiter='\n'):
-        items = parse(prev_results).values
-        k, acc, *args = commands
-
-        for item in items:
-            command = Terms([k, acc] + args)
-            acc = self.run_commands(command, item, run=True)
-
-            if str(acc).strip() == '' and self._last_results:
-                acc = self._last_results[-1]
-                self.env[LAST_RESULTS] = []
-
-        return acc
-
-    def do_flatten(self, args: str) -> str:
-        """Convert a space-separated string to a newline-separates string.
-        """
-        return '\n'.join(args.split(' '))
-
     def do_int(self, args: str) -> str:
         self._save_result(int(args))
         return ''
@@ -319,95 +176,115 @@ class BaseShell(Cmd):
         return ''
 
     def do_not(self, args: str) -> str:
+        # TODO mv to shell.model
         return FALSE if to_bool(args) else TRUE
+
+    def do_env(self, keys: str):
+        """Retrieve environment variables.
+        Return all variables if no key is given.
+        """
+        if not keys:
+            return filter_private_keys(self.env.asdict())
+
+        try:
+            return {k: self.env[k] for k in keys.split()}
+        except KeyError:
+            log('Invalid key')
+
+    def do_save(self, _):
+        """Save the current session.
+        """
+        self.save_session()
+
+    def do_reload(self, _):
+        """Reload the current session.
+        """
+        self.try_load_session()
+
+    def do_undo(self, _):
+        """Undo the previous command
+        """
+        if not self.lastcmd:
+            return
+
+        f = self.lastcmd.split()[0]
+
+        method = f'undo_{f}'
+        if has_method(self, f'undo_{f}'):
+            return getattr(self, method)()
+
+        raise NotImplementedError()
+
+    def last_method(self):
+        """Find the method corresponding to the last command run in `shell`.
+        It has the form: do_{cmd}
+
+        Return a the last method if it exists and None otherwise.
+        """
+        # TODO integrate this into Shell and store the last succesful cmd
+
+        if not self.lastcmd:
+            return
+
+        cmd = self.lastcmd.split()[0]
+        return BaseShell.get_method(cmd)
+
+    @staticmethod
+    def get_method(method_suffix: str):
+        method_name = f'do_{method_suffix}'
+        if not has_method(BaseShell, method_name):
+            return
+
+        method = getattr(BaseShell, method_name)
+
+        if isinstance(method, Function):
+            return method.func
+
+        return method
+
+    def add_special_function(self, char: str, method: Command):
+        # TODO merge this with self.add_functions
+        if char in delimiters.all or char in '!?':
+            raise ShellError(f'Char {char} is already in use.')
+
+        self.add_functions({char: method}, CHAR)
+
+    def run_special_function(self, k: str, args):
+        return self.function_groups[CHAR][k](*args)
+
+    def run_hidden_function(self, k: str, args):
+        for group in self.function_groups.values():
+            if k in group:
+                return group[k](*args)
+
+        raise ShellError(f'Unknown function: {k}')
+
+    def add_functions(self, functions: Dict[str, Function], group_key=None):
+        """Add functions to this instance at runtime.
+        Use a key to select a group of functions
+        """
+        if group_key is None:
+            group_key = default_function_group_key
+
+        for key, func in functions.items():
+            if hasattr(self, f'do_{key}'):
+                logging.debug('Warning: overriding self.do_{key}')
+
+            self.function_groups[group_key][key] = func
+
+    def remove_functions(self, group_key=None):
+        """Remove functions to this instance at runtime.
+        Use a key to select a group of functions
+        """
+        if group_key is None:
+            group_key = default_function_group_key
+
+        if group_key in self.function_groups:
+            del self.function_groups[group_key]
 
     ############################################################################
     # Overrides
     ############################################################################
-
-    def onecmd_inner(self, lines: str):
-        if not self.use_model:
-            return super().onecmd_inner(lines)
-
-        ast = parse(lines)
-        if ast is None:
-            raise ShellError('Invalid syntax: AST is empty')
-
-        try:
-            self.run_commands(ast, '', run=True)
-
-        except ShellPipeError as e:
-            if self.ignore_invalid_syntax:
-                log(e)
-                return
-
-            raise ShellError(e)
-
-        except subprocess.CalledProcessError as e:
-            returncode, stderr = e.args
-            log(f'Shell exited with {returncode}: {stderr}')
-
-            if self.ignore_invalid_syntax:
-                return
-
-            raise ShellError(str(e))
-
-    def run_commands(self, ast: Tuple, prev_result='', run=False):
-        if isinstance(ast, Term):
-            return ast.run(prev_result, shell=self, lazy=not run)
-
-        elif isinstance(ast, str):
-            return self.run_commands(Term(ast), prev_result, run=run)
-
-        done = self._handle_define_function(ast)
-        if done:
-            return
-
-        if not isinstance(ast, ElseCondition) and \
-                not isinstance(ast, Indent) and \
-                not isinstance(ast, Lines):
-            try:
-                handle_prev_then_else_statements(self)
-            except Abort:
-                return prev_result
-
-        if isinstance(ast, Node):
-            return ast.run(prev_result, shell=self, lazy=not run)
-        else:
-            raise NotImplementedError()
-
-    def _handle_define_function(self, ast: Node) -> bool:
-        if DEFINE_FUNCTION not in self.locals:
-            return False
-
-        # self._extend_inline_function_definition(line)
-        f = self.locals[DEFINE_FUNCTION]
-
-        if isinstance(ast, Indent):
-            # TODO compare indent width
-            width = ast.indent
-            if ast.data is None:
-                return True
-
-            if f.line_indent is None:
-                f.line_indent = width
-
-            if width >= f.line_indent:
-                self.locals[DEFINE_FUNCTION].inner.append(ast)
-                return True
-
-            self._finalize_define_function(f)
-
-        elif not isinstance(ast, Lines):
-            # TODO this will only be triggered after a non-Word command
-            self._finalize_define_function(f)
-
-        return False
-
-    def _finalize_define_function(self, f):
-        self.env[f.func_name] = f
-        self.locals.rm(DEFINE_FUNCTION)
-        self.prompt = default_prompt
 
     def postcmd(self, stop, _):
         """Display the shell_ready_signal to indicate termination to a parent process.
@@ -421,56 +298,6 @@ class BaseShell(Cmd):
 
         print_shell_ready_signal()
         return stop
-
-    ############################################################################
-    # Environment Variables
-    ############################################################################
-
-    def set_env_variable(self, k: str, *values: str):
-        """Set the variable `k` to `values`
-        """
-        if not is_valid_method_name(k):
-            raise ShellError(f'Invalid variable name format: {k}')
-
-        log(f'set {k}')
-        self.env[k] = ' '.join(values)
-        return k
-
-    def set_env_variables(self, keys: Union[str, List[str]], result: str):
-        """Set the variables `keys` to the values in result.
-        """
-        if result is None:
-            raise ShellError(f'Missing return value in assignment: {keys}')
-
-        if isinstance(keys, str) or isinstance(keys, Term):
-            keys = keys.split(' ')
-
-        try:
-            if len(result) == len(keys):
-                self.env.update(items=zip(keys, result))
-                return
-        except TypeError:
-            pass
-
-        if len(keys) == 1:
-            if isinstance(result, list):
-                result = ' '.join(quote_all(result))
-            self.env[keys[0]] = result
-        elif isinstance(result, str) or isinstance(result, Term):
-            lines = result.split('\n')
-            terms = result.split(' ')
-            if len(lines) == len(keys):
-                self.env.update(items=zip(keys, lines))
-
-            elif len(terms) == len(keys):
-                self.env.update(items=zip(keys, terms))
-
-            elif result == '':
-                self.env.update(items=zip(keys, repeat('')))
-
-        else:
-            raise ShellError(
-                f'Cannot assign values to all keys: {" ".join(keys)}')
 
     ############################################################################
     # Persistency: Save/load sessions to disk
@@ -532,13 +359,6 @@ class BaseShell(Cmd):
         self.env.update(env)
 
         self.load_session_posthook()
-
-    ############################################################################
-    # Inline & Multiline Functions
-    ############################################################################
-
-    def parse(self, results: str):
-        return parse(results)
 
 
 def is_public(key: str) -> bool:
