+++ src/mash/shell/function.py
@@ -1,3 +1,4 @@
+from collections import defaultdict
 from copy import deepcopy
 from dataclasses import dataclass, field
 from types import TracebackType
@@ -14,6 +15,10 @@ exception_hint = '(run `E` for details)'
 # global cache: sys.last_value and sys.last_traceback don't store exceptions raised in cmd.Cmd
 last_exception: Exception = None
 last_traceback: TracebackType = None
+
+
+LAST_RESULTS = '_last_results'
+LAST_RESULTS_INDEX = '_last_results_index'
 
 
 class ShellFunction:
@@ -64,3 +69,7 @@ class InlineFunction:
     multiline: bool = False
     line_indent: str = None
     inner: List[str] = field(default_factory=list)
+
+
+def scope() -> dict:
+    return defaultdict(dict)
