+++ test/shell/test_lib.py
@@ -3,7 +3,8 @@ from pytest import raises
 
 from mash import io_util
 from mash.shell import ShellError
-from mash.shell.shell import Shell, run_command, run_commands_from_file
+from mash.shell.shell import Shell
+from src.mash.shell.cmd2 import run_command, run_commands_from_file
 
 
 # Beware of the trailing space
