+++ test/shell/test_shell_multiline_functions.py
@@ -1,6 +1,7 @@
 from mash import io_util
 from mash.shell.delimiters import TRUE
-from mash.shell.shell import Shell, run_command
+from mash.shell.shell import Shell
+from src.mash.shell.cmd2 import run_command
 from mash.util import use_recursion_limit
